adding core repository with model methods + refactoring ClientRepository.php

+++ app/Http/Controllers/Client/ClientController.php
@@ -8,9 +8,12 @@ use App\Enums\ClientStatusEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Client\ClientStoreRequest;
 use App\Http\Requests\Client\ClientUpdateRequest;
-use App\Models\Client;
+use App\Repositories\Client\ClientRepository;
 use App\Services\ClientService;
 use Exception;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\File;
@@ -18,11 +21,18 @@ use Illuminate\Support\Facades\File;
 class ClientController extends Controller
 {
     /**
+     * @param ClientRepository $clientRepository
+     */
+    public function __construct(readonly ClientRepository $clientRepository)
+    {
+    }
+
+    /**
      * Display a listing of the resource.
      */
-    public function index(): \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application
+    public function index(): Application|Factory|View|\Illuminate\Foundation\Application
     {
-        $clients = Client::latest()->paginate(10);
+        $clients = $this->clientRepository->getAllWithPaginate(5);
 
         return view('client.index',compact('clients'))
             ->with('i', (request()->input('page', 1) - 1) * 5);
@@ -31,7 +41,7 @@ class ClientController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create(): \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application
+    public function create(): Application|Factory|View|\Illuminate\Foundation\Application
     {
         $id_status_new = ClientStatusEnum::NEW->value;
 
@@ -66,12 +76,16 @@ class ClientController extends Controller
      * Display the specified resource.
      *
      * @param $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|JsonResponse
+     * @return Application|Factory|View|\Illuminate\Foundation\Application|JsonResponse
      */
-    public function show($id): \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|JsonResponse
+    public function show($id): Application|Factory|View|\Illuminate\Foundation\Application|JsonResponse
     {
         try {
-            $client = Client::query()->findOrFail($id);
+            $client = $this->clientRepository->getEditModel($id);
+
+            if (empty($client)) {
+                abort(404);
+            }
 
             $clientPhotoPath = '';
 
@@ -89,12 +103,12 @@ class ClientController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|JsonResponse
+     * @return Application|Factory|View|\Illuminate\Foundation\Application|JsonResponse
      */
-    public function edit($id): \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|JsonResponse
+    public function edit($id): Application|Factory|View|\Illuminate\Foundation\Application|JsonResponse
     {
         try {
-            $client = Client::query()->findOrFail($id);
+            $client = $this->clientRepository->getEditModel($id);
 
             $clientPhotoPath = '';
 
@@ -123,13 +137,13 @@ class ClientController extends Controller
         try {
             $clientUpdateDTO = new ClientUpdateDTO($clientUpdateRequest, (int) $id);
 
-            $updateClient = $clientService->processUpdate($clientUpdateDTO);
+            $updateClient = $clientService->processUpdate($clientUpdateDTO, $this->clientRepository);
 
             if ($updateClient) {
-                return redirect()->route('clients.index')->with('success','Данные клиента успешно обновлены.');
+                return redirect()->route('clients.index')->with('success', sprintf('Данные клиента #%s успешно обновлены.', $id));
             }
 
-            return redirect()->route('clients.edit')->with('error','Ошибка! Данные клиента не обновлены.');
+            return back()->with('error', sprintf('Ошибка! Данные клиента #%s не обновлены.', $id));
         } catch (Exception $exception) {
             return response()->json(['error' => $exception->getMessage()], 401);
         }
@@ -144,7 +158,11 @@ class ClientController extends Controller
     public function destroy($id): RedirectResponse|JsonResponse
     {
         try {
-            $client = Client::query()->findOrFail($id);
+            $client = $this->clientRepository->getEditModel($id);
+
+            if (empty($client)) {
+                return back()->with('error', sprintf('Ошибка! Клиент #%s не удалён.', $id));
+            }
 
             $deleteClient = $client->delete();
 
@@ -152,7 +170,7 @@ class ClientController extends Controller
                 return redirect()->route('clients.index')->with('success','Клиент успешно удалён.');
             }
 
-            return redirect()->route('clients.index')->with('error','Ошибка! Клиент не удалён.');
+            return back()->with('error', sprintf('Ошибка! Клиент #%s не удалён.', $id));
         } catch (Exception $exception) {
             return response()->json(['error' => $exception->getMessage()], 401);
         }
